Add unit test
