+++ pom.xml
@@ -47,7 +47,6 @@
         <dependency>
 		    <groupId>org.junit.jupiter</groupId>
 		    <artifactId>junit-jupiter-api</artifactId>
-		    <version>5.1.0</version>
 		    <scope>test</scope>
 		</dependency>
         <dependency>
